@@ -1,14 +1,14 @@
 {
   "name": "react-step-wizard",
-  "version": "0.8.1",
+  "version": "2.0.3",
   "description": "Paginated step wizard",
   "scripts": {
-    "test": "eslint lib/ spec/ && ./node_modules/karma/bin/karma start karma.conf.js",
-    "watch-test": "./node_modules/karma/bin/karma start karma.conf.js --auto-watch --no-single-run",
+    "test": "eslint lib/ spec/ && karma start karma.conf.js",
+    "watch-test": "karma start karma.conf.js --auto-watch --no-single-run",
     "build": "webpack -p",
-    "dev": "webpack-dev-server --inline --hot"
+    "dev": "webpack-dev-server --inline --hot --config webpack.demo.config.js"
   },
-  "main": "./lib/index.js",
+  "main": "./dist/react-step-wizard.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/BI/react-step-wizard.git"
@@ -25,30 +25,38 @@
   },
   "homepage": "https://github.com/BI/react-step-wizard",
   "dependencies": {
-    "babel-runtime": "~5.4.3",
-    "classnames": "^1.2.2",
-    "react": "~0.13.0",
-    "react-event-listener": "0.0.2",
+    "classnames": "^2.2.5",
+    "react-event-listener": "^0.5.3",
     "react-pill-selector": "git://github.com/BI/react-pill-selector"
   },
+  "peerDependencies": {
+    "react": ">=15.1.0"
+  },
   "devDependencies": {
-    "babel-core": "~5.4.3",
-    "babel-eslint": "^3.1.4",
-    "babel-loader": "~5.1.2",
-    "css-loader": "~0.12.1",
-    "es5-shim": "~4.1.0",
-    "eslint": "^0.21.1",
-    "eslint-plugin-react": "^2.3.0",
-    "jasmine": "~2.2.1",
-    "jasmine-core": "~2.2.0",
-    "karma": "~0.12.31",
-    "karma-jasmine": "~0.3.5",
-    "karma-phantomjs-launcher": "~0.1.4",
-    "karma-sourcemap-loader": "~0.3.4",
-    "karma-webpack": "~1.5.0",
-    "sass-loader": "~1.0.2",
-    "style-loader": "~0.12.2",
-    "webpack": "^1.10.1",
-    "webpack-dev-server": "^1.10.1"
+    "babel-core": "~6.9.0",
+    "babel-eslint": "^6.0.4",
+    "babel-loader": "~6.2.4",
+    "babel-preset-react": "^6.5.0",
+    "create-react-class": "^15.6.3",
+    "css-loader": "~0.23.1",
+    "es5-shim": "~4.5.8",
+    "eslint": "^2.10.2",
+    "eslint-plugin-react": "^5.1.1",
+    "jasmine": "~2.4.1",
+    "jasmine-core": "~2.4.1",
+    "karma": "~0.13.22",
+    "karma-jasmine": "~1.0.2",
+    "karma-phantomjs-launcher": "~1.0.0",
+    "karma-sourcemap-loader": "~0.3.7",
+    "karma-webpack": "~1.7.0",
+    "node-sass": "^3.7.0",
+    "phantomjs-prebuilt": "^2.1.7",
+    "prop-types": "^15.6.1",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
+    "sass-loader": "~3.2.0",
+    "style-loader": "~0.13.1",
+    "webpack": "^1.13.1",
+    "webpack-dev-server": "^1.14.1"
   }
 }
